feat: Adicionar mídia em uma lista pela homepage
