+++ frontend/src/components/Media/viewMedia.module.css
@@ -4,7 +4,6 @@
   padding: 10px;
   border-radius: 8px;
   max-width: 1000px;
-  margin: auto;
   flex-direction: column;
 }
 
@@ -150,9 +149,6 @@
   background-color: #464662;
 }
 
-.castSection {
-  margin-top: 40px;
-}
 .mainSection {
   margin-top: 40px;
   display: flex;
@@ -226,43 +222,6 @@
   color: #727278;
 }
 
-.columnDetails {
+.column {
   position: relative;
-  z-index: 5;
-}
-
-.config {
-  position: absolute;
-  top: 270px;
-  right: 55px;
-  background-color: var(--gray01);
-  width: 220px;
-  display: flex;
-  flex-direction: column;
-  padding: 5px;
-  border-radius: 5px;
-  opacity: 0;
-  transition: all 0.4s ease-in-out;
-  z-index: 10;
-}
-
-.dropdown-anime {
-  opacity: 1;
-}
-
-.config li btnAddLista span {
-  color: var(--white);
-}
-
-.config .btnAddLista:hover {
-  background-color: var(--gray03);
-}
-
-.btnAddLista {
-  border: none;
-  background: transparent;
-  width: 100%;
-  padding: 10px;
-  cursor: pointer;
-  color: var(--white);
 }
