+++ frontend/src/components/Media/viewMedia.tsx
@@ -1,17 +1,11 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import styles from "./ViewMedia.module.css";
 import type { Media } from "../../features/media/media.type";
-import { baseURL } from "../../api";
 import { CreateReview } from "../review/CreateReview";
 import ListReview from "../review/ListReview";
 import { closeModal } from "../../features/ui/uiSlice";
 import { useAppDispatch, useAppSelector } from "../../app/store";
-import {
-  createListaAction,
-  updateListaAction,
-} from "../../features/lista/listaSlice";
-import { Link, useNavigate } from "react-router-dom";
-import { MidiaList } from "../../features/lista/lista.type";
+import AddListasDropdown from "../list/AddListasDropdown";
 
 const getBadgeClass = (classification) => {
   switch (classification) {
@@ -59,55 +53,13 @@ const ViewMedia: React.FC<Media> = ({
     setShowReview((showReview) => !showReview);
   };
   const dispatch = useAppDispatch();
-  const navigate = useNavigate();
+  const { media } = useAppSelector((state) => state.media);
 
   useEffect(() => {
     dispatch(closeModal());
   }, [dispatch]);
-  const { user } = useAppSelector((state) => state.user);
-  const { media } = useAppSelector((state) => state.media);
-  const { listas } = useAppSelector((state) => state.user.user);
+
   const [showListas, setShowListas] = React.useState(false);
-
-  function addToLista(id: number) {
-    const lista = listas.find((lista) => lista.id == id);
-    const data = {
-      id: lista.id,
-      name: lista.name,
-      description: lista.description,
-      midias: [...lista.midias.map((midia) => midia.id), media?.id],
-    };
-    dispatch(
-      updateListaAction(
-        data,
-        () => {
-          navigate(`/list/${lista.id}`);
-        },
-        () => {}
-      )
-    );
-  }
-
-  function handleCreateList() {
-    console.log(user?.listas);
-    const newLista = {
-      name: "Minha Lista",
-      description: "Adicione uma descrição...",
-      midias: [],
-      user: user?.id,
-    };
-    dispatch(
-      createListaAction(
-        newLista,
-        () => {
-          navigate("/list");
-        },
-        () => {
-          console.log("erro");
-        }
-      )
-    );
-  }
 
   return (
     <>
@@ -116,7 +68,7 @@ const ViewMedia: React.FC<Media> = ({
           <div className={styles.container}>
             <div className={styles.columnPoster}>
               <div className={styles.poster}>
-                <img src={media?.banner} alt={`${title} Poster`} />
+                <img src={banner} alt={`${title} Poster`} />
               </div>
             </div>
             <div className={styles.columnDetails}>
@@ -185,39 +137,14 @@ const ViewMedia: React.FC<Media> = ({
                       >
                         <i className="fa-solid fa-plus"></i> Adicionar na Lista
                       </button>
+                      <AddListasDropdown
+                        showListas={showListas}
+                        midiaId={media.id}
+                      />
                     </div>
                   </div>
                 </div>
               </div>
-              <ul
-                className={
-                  showListas
-                    ? `${styles.config} ${styles["dropdown-anime"]}`
-                    : `${styles.config}`
-                }
-              >
-                {listas.length !== 0 ? (
-                  listas.map((lista) => (
-                    <li key={lista.id}>
-                      <button
-                        className={styles.btnAddLista}
-                        onClick={() => addToLista(lista.id)}
-                      >
-                        <span>{lista.name}</span>
-                      </button>
-                    </li>
-                  ))
-                ) : (
-                  <li>
-                    <button
-                      className={styles.btnAddLista}
-                      onClick={handleCreateList}
-                    >
-                      <span>Criar lista</span>
-                    </button>
-                  </li>
-                )}
-              </ul>
             </div>
           </div>
         </div>
