+++ frontend/src/components/list/AddListasDropdown.module.css
@@ -1,0 +1,60 @@
+.config {
+  position: absolute;
+  top: 57px;
+  right: 0;
+  background-color: var(--gray01);
+  width: 220px;
+  height: calc(320px - 57px);
+  display: none;
+  flex-direction: column;
+  padding: 15px;
+  padding-right: 5px;
+  border-radius: 5px;
+  opacity: 0;
+  transition: all 0.4s ease-in-out;
+  z-index: 10;
+  overflow-y: scroll;
+}
+
+.config::-webkit-scrollbar {
+  width: 10px;
+  height: 10px;
+}
+
+.config::-webkit-scrollbar-track {
+  background: var(--gray01);
+  border-radius: 8px;
+}
+
+.config::-webkit-scrollbar-thumb {
+  background-color: #f1f1f1;
+  border-radius: 10px;
+  border: 3px solid var(--gray01);
+}
+
+.config::-webkit-scrollbar-thumb:hover {
+  background-color: #f1f1f1;
+}
+
+.dropdown-anime {
+  opacity: 1;
+  display: flex;
+}
+
+.config li btnAddLista span {
+  color: var(--white);
+}
+
+.config .btnAddLista:hover {
+  background-color: var(--gray03);
+}
+
+.btnAddLista {
+  border: none;
+  background: transparent;
+  width: 100%;
+  padding: 10px;
+  cursor: pointer;
+  color: var(--white);
+  border-radius: 3px;
+}
