+++ frontend/src/components/list/AllListsView.module.css
@@ -1,7 +1,6 @@
 .list-container {
   display: grid;
   grid-template-columns: 1fr 210px;
-  width: 100%;
   margin: 0 10%;
   gap: 60px;
 }
