+++ frontend/src/features/lista/lista.type.ts
@@ -1,11 +1,12 @@
 export interface Lista {
   name: string;
   description: string;
-  midias: MediaList[];
+  midias: (number | undefined)[];
 }
 
 export interface MidiaList {
   id: number;
   title: string;
   banner: string;
+  media_type: string;
 }
